Move to "simple" package.
The "simple" package namespace is used to represent the type of API. Namely, this API provides a single unit of multiple independently versioned schema entities (metadata, basic, and advanced). Other potential APIs for accessing the same metadata may be used that can adapt to multiple versions simultaneously. The term "simple" is used because "basic" has semantic meaning within the data and schema itself.